Header separation and linux compilation
